ALV Torrance lot: The Form now properly submits (The headache...)
Couldn't believe how long it took me to figure out that the culprit was: e.preventDefault(). It was literally doing its job and preventing the form submission from occurring smh.

ALSO, I did use Postman (the utility to send artificial requests) to send a 'post' request to the corresponding URL and it worked.....but that was super misleading. I thought something else was the problem but Postman bypasses the whole 'form' in the first place.

ALSO, every element of the form NEEDS to have a 'name' attribute....yes even the submit button. I know because I just added the "name" attribute to the button because the 'post' request was not sending my values. It was sending blanks.